# Conflicts:
#	Makefile
#	architecture/max-msp/ui.js
#	tests/soundfile/t5.dsp

+++ compiler/generator/instructions.hh
@@ -1089,7 +1089,7 @@ struct BlockInst : public StatementInst
         }
     }
     
-    size_t size() { return fCode.size(); }
+    int size() { return int(fCode.size()); }
 
     bool hasReturn();
     ValueInst* getReturnValue();
